pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "prometheus-ptg"
dynamic = ["version"]
description = "Prometheus — Teach-To-Grow 种子基因编辑器，史诗编史官系统"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Audrey · 001X", email = "audrey@ptg.dev" }]
license = { text = "MIT" }
keywords = [
    "ai-agent",
    "llm",
    "claude",
    "prometheus",
    "seed-gene-editor",
    "chronicler",
    "teach-to-grow",
    "multi-platform",
    "gateway",
    "skills",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "openai>=2.21.0,<3",
    "anthropic>=0.39.0,<1",
    "python-dotenv>=1.2.1,<2",
    "fire>=0.7.1,<1",
    "httpx[socks]>=0.28.1,<1",
    "rich>=14.3.3,<15",
    "tenacity>=9.1.4,<10",
    "pyyaml>=6.0.2,<7",
    "requests>=2.33.0,<3",
    "jinja2>=3.1.5,<4",
    "pydantic>=2.12.5,<3",
    "prompt_toolkit>=3.0.52,<4",
    "lark-oapi>=1.5.3,<2",
]

[project.optional-dependencies]
dev = ["debugpy>=1.8.0,<2", "pytest>=9.0.2,<10", "pytest-asyncio>=1.3.0,<2", "ruff", "mypy>=1.14.0,<2", "import-linter>=2.0,<3"]
mcp = ["mcp>=1.2.0,<2"]
web = ["fastapi>=0.104.0,<1", "uvicorn[standard]>=0.24.0,<1"]
file_watch = ["watchdog>=4.0.0,<5"]
vector_memory = ["numpy>=1.24.0,<3"]

# 消息频道支持（可选安装）
messaging = ["python-telegram-bot[webhooks]>=22.6,<23", "discord.py[voice]>=2.7.1,<3", "aiohttp>=3.13.3,<4", "slack-bolt>=1.18.0,<2", "slack-sdk>=3.27.0,<4"]
cron = ["croniter>=6.0.0,<7"]
dingtalk = ["dingtalk-stream>=0.20,<1", "alibabacloud-dingtalk>=2.0.0"]
feishu = ["lark-oapi>=1.5.3,<2"]

# Skills 和 Plugins（可选安装）
skills = ["prometheus-ptg[cron]", "prometheus-ptg[mcp]"]
plugins = ["prometheus-ptg[skills]"]

termux = [
    # Tested Android / Termux path: keeps the core CLI feature-rich while
    # avoiding extras that currently depend on non-Android wheels.
    "python-telegram-bot[webhooks]>=22.6,<23",
    "prometheus-ptg[cron]",
    "prometheus-ptg[mcp]",
]

# 完整安装（包含向量记忆等）
full = [
    "prometheus-ptg[vector_memory]",
    "prometheus-ptg[mcp]",
    "prometheus-ptg[web]",
    "prometheus-ptg[file_watch]",
    "prometheus-ptg[messaging]",
    "prometheus-ptg[cron]",
    "prometheus-ptg[dingtalk]",
    "prometheus-ptg[feishu]",
    "prometheus-ptg[skills]",
    "prometheus-ptg[plugins]",
]

# 完整安装（可选分组）
all = [
  "prometheus-ptg[dev]",
  "prometheus-ptg[mcp]",
  "prometheus-ptg[web]",
  "prometheus-ptg[file_watch]",
  "prometheus-ptg[messaging]",
  "prometheus-ptg[cron]",
  "prometheus-ptg[dingtalk]",
  "prometheus-ptg[feishu]",
  "prometheus-ptg[skills]",
  "prometheus-ptg[plugins]",
]

[project.scripts]
ptg = "prometheus.cli.main:main"
prometheus = "prometheus.cli.main:main"

[project.urls]
Homepage = "https://github.com/audrey/prometheus-ptg"
Documentation = "https://github.com/audrey/prometheus-ptg#readme"
Repository = "https://github.com/audrey/prometheus-ptg"
Issues = "https://github.com/audrey/prometheus-ptg/issues"
Changelog = "https://github.com/audrey/prometheus-ptg/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
include = ["prometheus", "prometheus.*"]

[tool.setuptools.dynamic]
version = {attr = "prometheus.__version__"}

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_optional = true
ignore_missing_imports = true
exclude = ["tests/", "skills/", "plugins/"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
    "docs_internal",
    "scripts",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes (unused imports, undefined names, etc.)
    "I",      # isort (import ordering)
    "N",      # pep8-naming
    "UP",     # pyupgrade (modernize syntax)
    "B",      # flake8-bugbear (common bugs)
    "SIM",    # flake8-simplify (simplify code)
    "TCH",    # flake8-type-checking (move imports to TYPE_CHECKING)
    "RUF100", # unused noqa directives
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "SIM108", # use ternary operator (style preference)
]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["prometheus"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
